using stdlib randomness, refs #3277
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@25121 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBNodeCont.cpp
@@ -344,7 +344,7 @@ NBNodeCont::removeComponents(NBDistrictCont& dc, NBEdgeCont& ec, const int numKe
             }
         }
         components.insert(cIt, component);
-        if (components.size() > numKeep) {
+        if ((int)components.size() > numKeep) {
             toRemove.insert(components.back().begin(), components.back().end());
             components.pop_back();
         }
